# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zawadi HR & Payroll API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /timesheets:
    get:
      operationId: listTimesheets
      tags: [timesheets]
      summary: List timesheets for a period
      parameters:
        - name: period
          in: query
          schema:
            type: string
          description: "YYYY-MM"
      responses:
        "200":
          description: Timesheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimesheetRow"
    post:
      operationId: upsertTimesheet
      tags: [timesheets]
      summary: Upsert timesheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimesheetInput"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
    patch:
      operationId: approveTimesheet
      tags: [timesheets]
      summary: Approve timesheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimesheetApprovalInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave requests
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRow"
    post:
      operationId: createLeave
      tags: [leaves]
      summary: Submit leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
    patch:
      operationId: decideLeave
      tags: [leaves]
      summary: Approve or reject leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveDecisionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List active loans
      responses:
        "200":
          description: Loans with repayment ledger
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoanRow"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Issue a direct loan (HR-initiated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
...
  /portal/leave:
    get:
      operationId: listPortalLeave
      tags: [portal]
      summary: Employee portal – my leave requests
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createPortalLeave
      tags: [portal]
      summary: Employee portal – submit leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
...
  /portal/p9:
    get:
      operationId: getPortalP9
      tags: [portal]
      summary: Employee portal – P9 annual tax summary
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: P9 data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/P9Data"
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, orgId, empNo, firstName, lastName, email, position, employmentType, status, basicSalary, hireDate]
      properties:
        id:
          type: integer
        orgId:
          type: integer
        empNo:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        gender:
          type: string
        nationalId:
          type: ["string", "null"]
        kraPin:
          type: ["string", "null"]
        nssfNo:
          type: ["string", "null"]
        shifNo:
          type: ["string", "null"]
        payMethod:
          type: string
        bankName:
          type: ["string", "null"]
        bankAccount:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        position:
          type: string
        employmentType:
          type: string
        residentStatus:
          type: string
        disabilityExemption:
          type: boolean
        basicSalary:
          type: integer
        houseAllowance:
          type: integer
        transportAllowance:
          type: integer
        otherAllowance:
          type: integer
        nonCashBenefit:
          type: integer
        insurancePremium:
          type: integer
        pensionEmployee:
          type: integer
        pensionEmployer:
          type: integer
        mortgageInterest:
          type: integer
        helbMonthly:
          type: integer
        saccoMonthly:
          type: integer
        hireDate:
          type: string
        terminationDate:
          type: ["string", "null"]
        status:
          type: string
        leaveBalance:
          type: integer
        createdAt:
          type: string
...
    Timesheet:
      type: object
      required: [id, orgId, employeeId, period, daysWorked, normalHours, overtimeHours, holidayHours]
      properties:
        id:
          type: integer
        orgId:
          type: integer
        employeeId:
          type: integer
        period:
          type: string
        daysWorked:
          type: integer
        normalHours:
          type: integer
        overtimeHours:
          type: integer
        holidayHours:
          type: integer
        approvedBy:
          type: ["integer", "null"]
        approvedAt:
          type: ["string", "null"]
    TimesheetRow:
      type: object
      required: [timesheet, employee]
      properties:
        timesheet:
          $ref: "#/components/schemas/Timesheet"
        employee:
          $ref: "#/components/schemas/Employee"
    TimesheetInput:
      type: object
      required: [employeeId, period, daysWorked, normalHours, overtimeHours, holidayHours]
      properties:
        employeeId:
          type: integer
        period:
          type: string
        daysWorked:
          type: integer
        normalHours:
          type: integer
        overtimeHours:
          type: integer
        holidayHours:
          type: integer
    TimesheetApprovalInput:
      type: object
      required: [id]
      properties:
        id:
          type: integer
...
    LeaveRequest:
      type: object
      required: [id, orgId, employeeId, type, startDate, endDate, days, status, createdAt]
      properties:
        id:
          type: integer
        orgId:
          type: integer
        employeeId:
          type: integer
        type:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        days:
          type: integer
        reason:
          type: ["string", "null"]
        status:
          type: string
        decidedByUserId:
          type: ["integer", "null"]
        decidedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    LeaveRow:
      type: object
      required: [leave, employee]
      properties:
        leave:
          $ref: "#/components/schemas/LeaveRequest"
        employee:
          $ref: "#/components/schemas/Employee"
    LeaveInput:
      type: object
      required: [type, startDate, endDate]
      properties:
        employeeId:
          type: integer
        type:
          type: string
          enum: [annual, sick, unpaid, maternity, paternity, compassionate]
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
    LeaveDecisionInput:
      type: object
      required: [id, action]
      properties:
        id:
          type: integer
        action:
          type: string
          enum: [approve, reject]
        note:
          type: string
    Loan:
      type: object
      required: [id, orgId, employeeId, type, principal, balance, monthlyInstallment, interestRateBps, status, startDate, createdAt]
      properties:
        id:
          type: integer
        orgId:
          type: integer
        employeeId:
          type: integer
        type:
          type: string
        principal:
          type: integer
        balance:
          type: integer
        monthlyInstallment:
          type: integer
        interestRateBps:
          type: integer
        status:
          type: string
        startDate:
          type: string
        createdAt:
          type: string
    LoanRow:
      type: object
      required: [loan, employee, repayments]
      properties:
        loan:
          $ref: "#/components/schemas/Loan"
        employee:
          $ref: "#/components/schemas/Employee"
        repayments:
          type: array
          items:
            type: object
            additionalProperties: true
        fringeBenefit:
          type: object
          properties:
            monthlyBenefit:
              type: integer
            monthlyTax:
              type: integer
          required: [monthlyBenefit, monthlyTax]
    LoanInput:
      type: object
      required: [employeeId, type, principal, months]
      properties:
        employeeId:
          type: integer
        type:
          type: string
          enum: [company, sacco, advance, emergency]
        principal:
          type: number
        interestRateBps:
          type: integer
        months:
          type: integer
        startDate:
          type: string
...
    P9Data:
      type: object
      required: [year, employee, months, totals]
      properties:
        year:
          type: integer
        employee:
          $ref: "#/components/schemas/Employee"
        months:
          type: array
          items:
            type: object
            additionalProperties: true
        totals:
          type: object
          additionalProperties: true